+++ src/components/Login.js
@@ -7,15 +7,12 @@ import {
   updateProfile,
 } from "firebase/auth";
 import { auth } from "../utils/firebase";
-import { useNavigate } from "react-router-dom";
 import { useDispatch } from "react-redux";
 import { addUser } from "../utils/userSlice";
 
 const Login = () => {
   const [isSignInForm, setIsSignInForm] = useState(true);
   const [errorMesssage, setErrorMessage] = useState(null);
-  const navigate = useNavigate();
-
   const dispatch = useDispatch();
 
   const name = useRef(null);
@@ -24,9 +21,6 @@ const Login = () => {
 
   const handleButtonClick = () => {
     //Validate the form data
-
-    // console.log(email.current.value);
-    //console.log(password.current.value);
 
     const message = checkValidData(email.current.value, password.current.value);
 
@@ -64,9 +58,6 @@ const Login = () => {
                   photoURL: photoURL,
                 })
               );
-
-              console.log(user);
-              navigate("/browse");
             })
             .catch((error) => {
               // An error occurred
@@ -93,8 +84,6 @@ const Login = () => {
         .then((userCredential) => {
           // Signed in
           const user = userCredential.user;
-          console.log(user);
-          navigate("/browse");
           // ...
         })
         .catch((error) => {
